+++ hotel/src/main/java/com/mjc/hotel/review/dto/ReviewPhotoRequest.java
@@ -2,6 +2,6 @@ package com.mjc.hotel.review.dto;
 
 public record ReviewPhotoRequest(
         String photoPath,
-        Long photoOrder
+        Integer photoOrder
 ) {
 }
